feat(rpc): v04 removal related cleanup
