+++ crates/rpc/src/v02/method.rs
@@ -1,5 +1,4 @@
 mod block_hash_and_number;
-pub(crate) mod call;
 mod chain_id;
 mod get_block_transaction_count;
 mod get_class;
@@ -11,7 +10,7 @@ pub(crate) mod get_transaction_by_block_id_and_index;
 pub(crate) mod get_transaction_by_hash;
 
 pub(crate) use block_hash_and_number::{block_hash_and_number, block_number};
-pub(crate) use call::call;
+
 pub(crate) use chain_id::chain_id;
 pub(crate) use get_block_transaction_count::get_block_transaction_count;
 pub(crate) use get_class::get_class;
