+++ crates/rpc/src/v02/types.rs
@@ -134,7 +134,9 @@ pub mod request {
     use pathfinder_crypto::hash::{HashChain, PoseidonHasher};
     use serde::Deserialize;
     use serde_with::serde_as;
-    use starknet_gateway_types::transaction_hash::compute_txn_hash;
+    use starknet_gateway_types::{
+        reply::transaction::EntryPointType, transaction_hash::compute_txn_hash,
+    };
 
     /// "Broadcasted" L2 transaction in requests the RPC API.
     ///
@@ -791,6 +793,171 @@ pub mod request {
                 self.fee_data_availability_mode.into(),
                 self.resource_bounds.into(),
             )
+        }
+    }
+
+    impl From<BroadcastedTransaction> for starknet_gateway_types::reply::transaction::Transaction {
+        fn from(value: BroadcastedTransaction) -> Self {
+            use starknet_gateway_types::reply::transaction as gateway;
+
+            match value {
+                BroadcastedTransaction::Declare(BroadcastedDeclareTransaction::V0(declare)) => {
+                    let class_hash = declare.contract_class.class_hash().unwrap().hash();
+                    let transaction_hash =
+                        declare.transaction_hash(ChainId::GOERLI_TESTNET, class_hash);
+                    gateway::Transaction::Declare(gateway::DeclareTransaction::V0(
+                        gateway::DeclareTransactionV0V1 {
+                            class_hash,
+                            max_fee: declare.max_fee,
+                            nonce: TransactionNonce::default(),
+                            sender_address: declare.sender_address,
+                            signature: declare.signature,
+                            transaction_hash,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::Declare(BroadcastedDeclareTransaction::V1(declare)) => {
+                    let class_hash = declare.contract_class.class_hash().unwrap().hash();
+                    let transaction_hash =
+                        declare.transaction_hash(ChainId::GOERLI_TESTNET, class_hash);
+                    gateway::Transaction::Declare(gateway::DeclareTransaction::V1(
+                        gateway::DeclareTransactionV0V1 {
+                            class_hash,
+                            max_fee: declare.max_fee,
+                            nonce: TransactionNonce::default(),
+                            sender_address: declare.sender_address,
+                            signature: declare.signature,
+                            transaction_hash,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::Declare(BroadcastedDeclareTransaction::V2(declare)) => {
+                    let class_hash = declare.contract_class.class_hash().unwrap().hash();
+                    let transaction_hash =
+                        declare.transaction_hash(ChainId::GOERLI_TESTNET, class_hash);
+                    gateway::Transaction::Declare(gateway::DeclareTransaction::V2(
+                        gateway::DeclareTransactionV2 {
+                            class_hash,
+                            max_fee: declare.max_fee,
+                            nonce: TransactionNonce::default(),
+                            sender_address: declare.sender_address,
+                            signature: declare.signature,
+                            transaction_hash,
+                            compiled_class_hash: declare.compiled_class_hash,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::Declare(BroadcastedDeclareTransaction::V3(declare)) => {
+                    let class_hash = declare.contract_class.class_hash().unwrap().hash();
+                    let transaction_hash =
+                        declare.transaction_hash(ChainId::GOERLI_TESTNET, class_hash);
+                    gateway::Transaction::Declare(gateway::DeclareTransaction::V3(
+                        gateway::DeclareTransactionV3 {
+                            class_hash,
+                            nonce: TransactionNonce::default(),
+                            sender_address: declare.sender_address,
+                            signature: declare.signature,
+                            transaction_hash,
+                            compiled_class_hash: declare.compiled_class_hash,
+                            nonce_data_availability_mode: declare
+                                .nonce_data_availability_mode
+                                .into(),
+                            fee_data_availability_mode: declare.fee_data_availability_mode.into(),
+                            resource_bounds: declare.resource_bounds.into(),
+                            tip: declare.tip,
+                            paymaster_data: declare.paymaster_data,
+                            account_deployment_data: declare.account_deployment_data,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::DeployAccount(
+                    BroadcastedDeployAccountTransaction::V0V1(deploy),
+                ) => gateway::Transaction::DeployAccount(gateway::DeployAccountTransaction::V0V1(
+                    gateway::DeployAccountTransactionV0V1 {
+                        contract_address: deploy.deployed_contract_address(),
+                        transaction_hash: deploy.transaction_hash(ChainId::GOERLI_TESTNET),
+                        max_fee: deploy.max_fee,
+                        version: deploy.version,
+                        signature: deploy.signature,
+                        nonce: deploy.nonce,
+                        contract_address_salt: deploy.contract_address_salt,
+                        constructor_calldata: deploy.constructor_calldata,
+                        class_hash: deploy.class_hash,
+                    },
+                )),
+                BroadcastedTransaction::DeployAccount(BroadcastedDeployAccountTransaction::V3(
+                    deploy,
+                )) => {
+                    let transaction_hash = deploy.transaction_hash(ChainId::GOERLI_TESTNET);
+
+                    gateway::Transaction::DeployAccount(gateway::DeployAccountTransaction::V3(
+                        gateway::DeployAccountTransactionV3 {
+                            version: deploy.version,
+                            class_hash: deploy.class_hash,
+                            nonce: deploy.nonce,
+                            sender_address: deploy.deployed_contract_address(),
+                            contract_address_salt: deploy.contract_address_salt,
+                            constructor_calldata: deploy.constructor_calldata,
+                            signature: deploy.signature,
+                            transaction_hash,
+                            nonce_data_availability_mode: deploy
+                                .nonce_data_availability_mode
+                                .into(),
+                            fee_data_availability_mode: deploy.fee_data_availability_mode.into(),
+                            resource_bounds: deploy.resource_bounds.into(),
+                            tip: deploy.tip,
+                            paymaster_data: deploy.paymaster_data,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::Invoke(BroadcastedInvokeTransaction::V0(invoke)) => {
+                    let transaction_hash = invoke.transaction_hash(ChainId::GOERLI_TESTNET);
+                    gateway::Transaction::Invoke(gateway::InvokeTransaction::V0(
+                        gateway::InvokeTransactionV0 {
+                            calldata: invoke.calldata,
+                            sender_address: invoke.contract_address,
+                            entry_point_type: Some(EntryPointType::External),
+                            entry_point_selector: invoke.entry_point_selector,
+                            max_fee: invoke.max_fee,
+                            signature: invoke.signature,
+                            transaction_hash,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::Invoke(BroadcastedInvokeTransaction::V1(invoke)) => {
+                    let transaction_hash = invoke.transaction_hash(ChainId::GOERLI_TESTNET);
+                    gateway::Transaction::Invoke(gateway::InvokeTransaction::V1(
+                        gateway::InvokeTransactionV1 {
+                            calldata: invoke.calldata,
+                            sender_address: invoke.sender_address,
+                            max_fee: invoke.max_fee,
+                            signature: invoke.signature,
+                            nonce: invoke.nonce,
+                            transaction_hash,
+                        },
+                    ))
+                }
+                BroadcastedTransaction::Invoke(BroadcastedInvokeTransaction::V3(invoke)) => {
+                    let transaction_hash = invoke.transaction_hash(ChainId::GOERLI_TESTNET);
+                    gateway::Transaction::Invoke(gateway::InvokeTransaction::V3(
+                        gateway::InvokeTransactionV3 {
+                            nonce: TransactionNonce::default(),
+                            sender_address: invoke.sender_address,
+                            signature: invoke.signature,
+                            transaction_hash,
+                            nonce_data_availability_mode: invoke
+                                .nonce_data_availability_mode
+                                .into(),
+                            fee_data_availability_mode: invoke.fee_data_availability_mode.into(),
+                            resource_bounds: invoke.resource_bounds.into(),
+                            tip: invoke.tip,
+                            paymaster_data: invoke.paymaster_data,
+                            calldata: invoke.calldata,
+                            account_deployment_data: invoke.account_deployment_data,
+                        },
+                    ))
+                }
+            }
         }
     }
