+++ crates/rpc/src/v03/method.rs
@@ -1,7 +1,5 @@
-pub(crate) mod estimate_fee;
 mod get_events;
 pub(crate) mod get_state_update;
 
-pub(crate) use estimate_fee::estimate_fee;
 pub(crate) use get_events::get_events;
 pub(crate) use get_state_update::get_state_update;
