+++ crates/rpc/src/v04.rs
@@ -1,4 +1,2 @@
 pub mod method;
 pub(crate) mod types;
-
-use crate::v04::method as v04_method;
