+++ crates/rpc/src/v04/method.rs
@@ -5,13 +5,10 @@ mod get_transaction_by_block_and_index;
 mod get_transaction_by_hash;
 pub(crate) mod simulate_transactions;
 mod syncing;
-mod trace_block_transactions;
 
 pub(crate) use add_declare_transaction::add_declare_transaction;
 pub(crate) use add_deploy_account_transaction::add_deploy_account_transaction;
 pub(crate) use add_invoke_transaction::add_invoke_transaction;
 pub(crate) use get_transaction_by_block_and_index::get_transaction_by_block_id_and_index;
 pub(crate) use get_transaction_by_hash::get_transaction_by_hash;
-// pub(super) use simulate_transactions::simulate_transactions;
 pub(crate) use syncing::syncing;
-// pub(super) use trace_block_transactions::trace_block_transactions;
